add hbs files and update routes

+++ src/controllers/index.controller.js
@@ -12,7 +12,7 @@ const getUsers = async (req, res) => {
   const response = await pool.query("SELECT * FROM users;");
   //console.log(response.rows);
   res.json(response.rows);
-  const Users = response.rows;
+  const { name, email } = response.rows;
   //res.send('All users')
   //res.render('/users/all-users')
 };
@@ -35,20 +35,32 @@ const createUser = async (req, res) => {
   } else {
     response = await pool.query("INSERT INTO users (name, email) VALUES ($1, $2)", [name, email]);
     console.log(name, email)
-    console.log("User created sucessfully");  
+    console.log("User created sucessfully");
     res.redirect('/users');
   }
 };
 
 const deleteUser = async (req, res) => {
-  const id = req.params.id;
-  const response = await pool.query("DELETE FROM users WHERE id = $1", [id]);
-  console.log(response);
-  res.send("User deleted sucessfully");
+  const id = req.body;
+  const errors = [];
+  if (!id) {
+    errors.push({ text: "Please Insert a Name" });
+  }
+  if (errors.length > 0) {
+    res.render("users/delete-user", {
+      errors,
+      id
+    });
+  } else {
+    const response = await pool.query("DELETE FROM users WHERE id = $1", [id]);
+    console.log(response);
+    res.send("User deleted sucessfully");
+  };
 };
+
 const updateUser = async (req, res) => {
   const { name, email } = req.body;
-  const id = req.params.id;
+  const id = req.body;
   const response = await pool.query(
     "UPDATE users SET name = $1, email = $2, WHERE id = $3",
     [name, email, id]
